Create comments seciton

+++ Controllers/IdeasController.cs
@@ -1,8 +1,8 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
-
-
+using Microsoft.EntityFrameworkCore;
 [Authorize]
 [ApiController]
 [Route("api/[controller]")]
@@ -18,39 +18,56 @@ public class IdeasController : ControllerBase
     [HttpGet]
     public IActionResult GetIdeas()
     {
-        var ideas = _context.Ideas.ToList();
+        var ideas = _context.Ideas
+    .Include(i => i.User).Select(i => new IdeaResponseDto
+{
+    Id = i.Id,
+    Title = i.Title,
+    Description = i.Description,
+    Category = i.Category,
+    CreatedAt = i.CreatedAt,
+    UserId = i.UserId,
+    UserName = i.User != null ? i.User.Name : null
+})
+
+
+    .ToList();
+
         return Ok(ideas);
     }
 
 
-[HttpPost]
-public IActionResult CreateIdea(CreateIdeaDto dto)
-{
-    var idea = new Idea
+    [HttpPost]
+    public IActionResult CreateIdea(CreateIdeaDto dto)
     {
-        Id = Guid.NewGuid(),
-        Title = dto.Title,
-        Description = dto.Description,
-        Category = dto.Category,
-        CreatedAt = DateTime.UtcNow
-    };
+        var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        var idea = new Idea
+        {
+            Id = Guid.NewGuid(),
+            Title = dto.Title,
+            Description = dto.Description,
+            Category = dto.Category,
+            CreatedAt = DateTime.UtcNow,
+            UserId = Guid.Parse(userId!)
+        };
 
-    _context.Ideas.Add(idea);
-    _context.SaveChanges();
 
-    return Ok(idea);
-}
+        _context.Ideas.Add(idea);
+        _context.SaveChanges();
 
-[HttpGet("{id}")]
-public IActionResult GetIdea(Guid id)
-{
-    var idea = _context.Ideas.Find(id);
+        return Ok(idea);
+    }
 
-    if (idea == null)
-        return NotFound();
+    [HttpGet("{id}")]
+    public IActionResult GetIdea(Guid id)
+    {
+        var idea = _context.Ideas.Find(id);
 
-    return Ok(idea);
-}
+        if (idea == null)
+            return NotFound();
+
+        return Ok(idea);
+    }
 
 
 [HttpDelete("{id}")]
@@ -61,6 +78,11 @@ public IActionResult DeleteIdea(Guid id)
     if (idea == null)
         return NotFound();
 
+    var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+    if (idea.UserId != Guid.Parse(userId!))
+        return Forbid();
+
     _context.Ideas.Remove(idea);
     _context.SaveChanges();
 
@@ -70,13 +92,18 @@ public IActionResult DeleteIdea(Guid id)
 
 
 
-[HttpPut("{id}")]
+ [HttpPut("{id}")]
 public IActionResult UpdateIdea(Guid id, UpdateIdeaDto dto)
 {
     var idea = _context.Ideas.Find(id);
 
     if (idea == null)
         return NotFound();
+
+    var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+    if (idea.UserId != Guid.Parse(userId!))
+        return Forbid();
 
     idea.Title = dto.Title;
     idea.Description = dto.Description;
